Prepare for rename to "Sense"
