+++ CONTRIBUTING.md
@@ -1,8 +1,8 @@
-# Contributing to 20bn-realtimenet
+# Contributing to Sense
 
 We welcome contributions from developers to improve our pipeline of tools. Additionally, we would love to showcase
 the amazing projects our community develops on our README.
 
 ## License
-By contributing to 20bn-realtimenet, you agree that the license found in the root directory in the file `LICENSE.md
+By contributing to Sense, you agree that the license found in the root directory in the file `LICENSE.md
 ` will extend to your contribution.
