+++ scripts/run_calorie_estimation.py
@@ -23,12 +23,12 @@ Options:
 """
 from docopt import docopt
 
-import realtimenet.display
-from realtimenet import engine
-from realtimenet import feature_extractors
-from realtimenet.controller import Controller
-from realtimenet.downstream_tasks import calorie_estimation
-from realtimenet.downstream_tasks.nn_utils import Pipe
+import sense.display
+from sense import engine
+from sense import feature_extractors
+from sense.controller import Controller
+from sense.downstream_tasks import calorie_estimation
+from sense.downstream_tasks.nn_utils import Pipe
 
 if __name__ == "__main__":
     # Parse arguments
@@ -68,11 +68,11 @@ if __name__ == "__main__":
     ]
 
     display_ops = [
-        realtimenet.display.DisplayFPS(expected_camera_fps=net.fps,
-                                       expected_inference_fps=net.fps / net.step_size),
-        realtimenet.display.DisplayDetailedMETandCalories(),
+        sense.display.DisplayFPS(expected_camera_fps=net.fps,
+                                 expected_inference_fps=net.fps / net.step_size),
+        sense.display.DisplayDetailedMETandCalories(),
     ]
-    display_results = realtimenet.display.DisplayResults(title=title, display_ops=display_ops)
+    display_results = sense.display.DisplayResults(title=title, display_ops=display_ops)
 
     # Run live inference
     controller = Controller(
