+++ scripts/run_fitness_rep_counter.py
@@ -17,14 +17,15 @@ Options:
 """
 from docopt import docopt
 
-import realtimenet.display
-from realtimenet import camera
-from realtimenet import engine
-from realtimenet import feature_extractors
-from realtimenet.controller import Controller
-from realtimenet.downstream_tasks.fitness_rep_counting import INT2LAB
-from realtimenet.downstream_tasks.nn_utils import Pipe, LogisticRegression
-from realtimenet.downstream_tasks.postprocess import PostprocessRepCounts, PostprocessClassificationOutput
+import sense.display
+from sense import engine
+from sense import feature_extractors
+from sense.controller import Controller
+from sense.downstream_tasks.fitness_rep_counting import INT2LAB
+from sense.downstream_tasks.nn_utils import LogisticRegression
+from sense.downstream_tasks.nn_utils import Pipe
+from sense.downstream_tasks.postprocess import PostprocessClassificationOutput
+from sense.downstream_tasks.postprocess import PostprocessRepCounts
 
 
 if __name__ == "__main__":
@@ -58,13 +59,13 @@ if __name__ == "__main__":
     ]
 
     display_ops = [
-        realtimenet.display.DisplayFPS(expected_camera_fps=net.fps,
-                                       expected_inference_fps=net.fps / net.step_size),
-        realtimenet.display.DisplayTopKClassificationOutputs(top_k=1, threshold=0.5),
-        realtimenet.display.DisplayRepCounts()
+        sense.display.DisplayFPS(expected_camera_fps=net.fps,
+                                 expected_inference_fps=net.fps / net.step_size),
+        sense.display.DisplayTopKClassificationOutputs(top_k=1, threshold=0.5),
+        sense.display.DisplayRepCounts()
     ]
-    display_results = realtimenet.display.DisplayResults(title=title, display_ops=display_ops,
-                                                         border_size=100)
+    display_results = sense.display.DisplayResults(title=title, display_ops=display_ops,
+                                                   border_size=100)
 
     # Run live inference
     controller = Controller(
