+++ sense/camera.py
