+++ sense/downstream_tasks/calorie_estimation/__init__.py
