+++ sense/downstream_tasks/fitness_activity_recognition/__init__.py
